feat: Add CONTRIBUTING guidelines for project collaboration
refactor: Update basic usage example to include MusicEngine namespace

refactor: Encapsulate Music struct within MusicEngine namespace

refactor: Refactor MusicManager class to encapsulate implementation details

feat: Introduce MusicPlayer class for handling music playback functionality

build: Add CMakeLists.txt for MusicManager module

refactor: Move MusicManager implementation to separate source file

feat: Implement MusicParser for extracting metadata and cover art from audio files

docs: Add readme for MusicManager module

feat: Initialize MusicPlayer module with basic structure
