+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,24 @@
+---
+name: Pull Request
+about: Submit a code contribution
+title: "[feat/fix/docs]: "
+labels: ''
+assignees: ''
+
+---
+
+### What is the purpose of this PR?
+
+*(Please provide a detailed description of the problem you are solving or the feature you are adding here.)*
+
+---
+
+### Checklist
+
+Before submitting this PR, please ensure you have completed all of the following steps:
+
+- [ ] I have carefully read and followed the project's [CONTRIBUTING.md](../docs/CONTRIBUTING.md) guide.
+- [ ] My code is based on the latest `develop` branch.
+- [ ] **I confirm that the target branch for this PR is `develop`, not `main`.**
+- [ ] I have added necessary tests for my changes.
+- [ ] Relevant documentation has been updated accordingly.
