+++ CMakeLists.txt
@@ -1,62 +1,49 @@
 cmake_minimum_required(VERSION 3.16)
 project(MusicManager)
 
-# Select whether to build example programs
+# Global settings and options
 option(MUSICMANAGER_BUILD_EXAMPLES "Build the example programs" ON)
-
-# Enable compile command to ease indexing with e.g. clangd
 set(CMAKE_EXPORT_COMPILE_COMMANDS TRUE)
-
-# Set C++ standard
 set(CMAKE_CXX_STANDARD 20)
 
+# Find dependencies
+set(SPDLOG_INSTALL ON)
 add_subdirectory(third_party/spdlog)
 find_package(PkgConfig REQUIRED)
-
 pkg_check_modules(FFMPEG REQUIRED IMPORTED_TARGET
     libavformat
     libavcodec
     libavutil
 )
 
-set(CMAKE_CXX_STANDARD 20)
+# Include subdirectories
+# This will execute src/CMakeLists.txt and define the MusicManager target
+add_subdirectory(src)
 
-add_library(${PROJECT_NAME} STATIC
-        src/MusicManager.cpp
-        src/MusicParser.cpp
-)
-
-target_include_directories(${PROJECT_NAME}
-    PUBLIC
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-        $<INSTALL_INTERFACE:include>
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/src
-)
-
-target_link_libraries(${PROJECT_NAME} PRIVATE
-        PkgConfig::FFMPEG
-)
+# Conditional linking logic
+if (${PROJECT_NAME} STREQUAL ${CMAKE_PROJECT_NAME})
+    # Scene 1: Building as the main project
+    # We link spdlog here and mark it as PRIVATE
+    # because for the final executable (like examples), it's an implementation detail
+    message(STATUS "MusicManager is main project. Linking spdlog as PRIVATE.")
+    target_link_libraries(MusicManager PRIVATE spdlog::spdlog)
+else()
+    # Scene 2: Included as a submodule
+    # We mark spdlog as PUBLIC.
+    # This means any parent target that links against MusicManager will automatically inherit the spdlog dependency.
+    # This is the key to "passing it up"!
+    message(STATUS "MusicManager is subproject. Linking spdlog as PUBLIC.")
+    target_link_libraries(MusicManager PUBLIC spdlog::spdlog)
+endif()
 
 if(MUSICMANAGER_BUILD_EXAMPLES)
-    # If the option is ON, add the examples directory as a subdirectory
     add_subdirectory(examples)
     message(STATUS "Building examples...")
 endif()
 
-
-# According to whether this is the main project or a subproject, set up install rules and linking
+# Installation rules remain at the top level, as this is a project-level decision
 if (${PROJECT_NAME} STREQUAL ${CMAKE_PROJECT_NAME})
     message(STATUS "MusicManager is being built as the main project. Install rules enabled.")
-
-
-    target_include_directories(${PROJECT_NAME} PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/third_party/spdlog/include
-    )
-
-    target_link_directories(${PROJECT_NAME} PRIVATE
-        spdlog::spdlog
-    )
 
     include(GNUInstallDirs)
     
@@ -64,7 +51,7 @@ if (${PROJECT_NAME} STREQUAL ${CMAKE_PROJECT_NAME})
         DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
     )
 
-    install(TARGETS ${PROJECT_NAME}
+    install(TARGETS MusicManager
         EXPORT MusicManagerTargets
         ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
         LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
@@ -91,13 +78,4 @@ if (${PROJECT_NAME} STREQUAL ${CMAKE_PROJECT_NAME})
     )
 else()
     message(STATUS "MusicManager is being built as a subproject. Install rules disabled.")
-
-    target_include_directories(${PROJECT_NAME} PUBLIC
-        ${CMAKE_CURRENT_SOURCE_DIR}/third_party/spdlog/include
-    )
-
-    target_link_directories(${PROJECT_NAME} PUBLIC
-        spdlog::spdlog
-    )
-
 endif()
