+++ examples/basic_usage.cpp
@@ -4,6 +4,8 @@
 #include "spdlog/sinks/stdout_color_sinks.h"
 #include "spdlog/spdlog.h"
 
+using namespace MusicEngine;
+
 std::vector<std::filesystem::path> music_dirs = {"../music_test", "/home/hao/音乐"}; // Adjusted path for convention
 
 // Helper function to print music information
