+++ src/MusicManager/readme.md
@@ -1,0 +1,2 @@
+### 模块说明
+负责管理曲库。检索、解析与管理音乐
