+++ src/MusicPlayer/MusicPlayer.cpp
@@ -1,0 +1,17 @@
+#include "MusicManager.h"
+#include "Music.h"
+#include <thread>
+#include <atomic>
+#include <mutex>
+#include <condition_variable>
+
+// FFmpeg headers
+extern "C" {
+#include <libavformat/avformat.h>
+#include <libavcodec/avcodec.h>
+}
+
+namespace MusicEngine {
+
+
+}
